requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
shap==0.44.1
fairlearn==0.10.0
scikit-learn==1.4.0
pandas==2.2.0
plotly==5.18.0
google-generativeai==0.4.1
firebase-admin==6.4.0
joblib==1.3.2
numpy==1.26.0
reportlab==4.1.0
langgraph==0.0.30
python-dotenv==1.0.0
python-multipart==0.0.9
